pyproject.toml
[tool.poetry]
name = "asciibee"
version = "0.1.0"
description = ""
authors = ["Jamey Nakama <nakamajamey@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Pillow = "^9.4"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
